requirements.txt
accelerate==0.33.0
beautifulsoup4==4.12.3
certifi
charset-normalizer
contourpy==1.1.1
cycler==0.12.1
diffusers==0.29.2
filelock
Flask==2.0.1
Flask-CORS==3.0.10
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.0
huggingface-hub==0.24.2
idna
importlib_metadata==8.2.0
importlib_resources==6.4.0
Jinja2
kiwisolver==1.4.5
MarkupSafe
matplotlib==3.7.5
networkx
numpy
opencv-python==4.10.0.84
packaging==24.1
Pillow
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML
regex==2024.7.24
requests
safetensors==0.4.3
scipy==1.10.1
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
sympy
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
typing_extensions
urllib3
wcwidth==0.2.13
zipp==3.19.2